requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.3
django-gravatar2==1.4.4
django-heroku==0.3.1
django-mdeditor==0.1.20
gunicorn==20.1.0
importlib-metadata==4.11.3
Markdown==3.3.6
Pillow==9.1.0
psycopg2==2.9.3
python-decouple==3.6
pytils==0.4.1
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.0.0
zipp==3.8.0
